Deleted added form in MyKitchen
